arreglando migraciones

+++ src/cloudinary/cloudinary.controller.ts
@@ -57,7 +57,7 @@ export class CloudinaryController {
                 public_id: file.public_id,
                 secure_url: file.secure_url,
                 resource_type: file.resource_type,
-                format: file.format,
+                format: file.format,z
                 duration: file.duration,
             }));
     }
@@ -93,6 +93,7 @@ import {
     BadRequestException,
     Delete,
     UseGuards,
+    InternalServerErrorException,
 } from "@nestjs/common";
 import { FileInterceptor } from "@nestjs/platform-express";
 import { Response } from "express";
@@ -114,19 +115,21 @@ export class CloudinaryController {
     @UseInterceptors(FileInterceptor("file", multerConfig))
     async upload(
         @UploadedFile() file: Express.Multer.File,
-        @Body() dto: UploadMediaDto,
+        @Body("name") name: string, // ← aquí
+        @Body("folder") folder?: string,
+        @Body("caption") caption?: string,
     ) {
         if (!file) throw new BadRequestException("No file provided");
-        const upload = await this.cloudinary.uploadMedia(file.path, dto);
+        if (!name) throw new BadRequestException("Debes indicar un nombre");
+
+        const upload = await this.cloudinary.uploadMedia(file.path, {
+            publicId: name, // ← PASAS tu nombre limpio
+            folder,
+            context: caption,
+        });
+
         await unlink(file.path);
-        return {
-            public_id: upload.public_id,
-            secure_url: upload.secure_url,
-            resource_type: upload.resource_type,
-            format: upload.format,
-            duration: upload.duration,
-            caption: dto.caption,
-        };
+        return { ...upload };
     }
 
     @Get("trainings")
